+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 - **new**:
     - implement `InitOneShotRandom` initialization strategy
     - make `SolverStep`, `InitializationStrategy` & `OptimizationStrategy` instances have a settable random seed
-    - add distinction between `internal` and `solver` (to be added) benchmarks in CLI
+    - add distinction between `internal` and `solver` benchmarks in CLI
 
 - **improved**:
     - add initialization of `SolverState` to the reported results